Split up script.cpp into five million files
