+++ src/game_api/script/usertypes/drops_lua.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <sol/forward.hpp>
+
+class ScriptImpl;
+
+namespace NDrops
+{
+    void register_usertypes(sol::state& lua);
+};
