+++ src/game_api/script/usertypes/entity_lua.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <sol/forward.hpp>
+
+namespace NEntity
+{
+void register_usertypes(sol::state& lua);
+};
