+++ src/game_api/script/usertypes/gui_lua.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <sol/forward.hpp>
+
+class ScriptImpl;
+
+namespace NGui
+{
+    void register_usertypes(sol::state& lua, ScriptImpl* script);
+};
