+++ src/game_api/script/usertypes/particles_lua.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <sol/forward.hpp>
+
+namespace NParticles
+{
+void register_usertypes(sol::state& lua);
+};
