+++ src/game_api/script/usertypes/player_lua.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <sol/forward.hpp>
+
+namespace NPlayer
+{
+void register_usertypes(sol::state& lua);
+};
