+++ src/game_api/script/usertypes/save_context.hpp
@@ -2,8 +2,9 @@
 
 #include <string>
 
-using MakeSavePathCallback = std::string (*)(std::string_view script_path, std::string_view script_name);
+#include <sol/forward.hpp>
 
+using MakeSavePathCallback = std::string (*)(std::string_view script_path, std::string_view script_name);
 void register_make_save_path(MakeSavePathCallback make_save_path_callback);
 
 class SaveContext
@@ -28,3 +29,8 @@ class LoadContext
     std::string_view script_path;
     std::string_view script_name;
 };
+
+namespace NSaveContext
+{
+void register_usertypes(sol::state& lua);
+}
