+++ src/game_api/script/usertypes/sound_lua.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <sol/forward.hpp>
+
+class ScriptImpl;
+class SoundManager;
+
+namespace NSound
+{
+void register_usertypes(sol::state& lua, ScriptImpl* script);
+};
