+++ src/game_api/script/usertypes/state_lua.cpp
@@ -1,0 +1,148 @@
+#include "state_lua.hpp"
+
+#include "state.hpp"
+
+#include <sol/sol.hpp>
+
+namespace NState
+{
+void register_usertypes(sol::state& lua)
+{
+    lua.new_usertype<StateMemory>(
+        "StateMemory",
+        "screen_last",
+        &StateMemory::screen_last,
+        "screen",
+        &StateMemory::screen,
+        "screen_next",
+        &StateMemory::screen_next,
+        "ingame",
+        &StateMemory::ingame,
+        "playing",
+        &StateMemory::playing,
+        "pause",
+        &StateMemory::pause,
+        "width",
+        &StateMemory::w,
+        "height",
+        &StateMemory::h,
+        "kali_favor",
+        &StateMemory::kali_favor,
+        "kali_status",
+        &StateMemory::kali_status,
+        "kali_altars_destroyed",
+        &StateMemory::kali_altars_destroyed,
+        "seed",
+        &StateMemory::seed,
+        "time_total",
+        &StateMemory::time_total,
+        "world",
+        &StateMemory::world,
+        "world_next",
+        &StateMemory::world_next,
+        "world_start",
+        &StateMemory::world_start,
+        "level",
+        &StateMemory::level,
+        "level_next",
+        &StateMemory::level_next,
+        "level_start",
+        &StateMemory::level_start,
+        "theme",
+        &StateMemory::theme,
+        "theme_next",
+        &StateMemory::theme_next,
+        "theme_start",
+        &StateMemory::theme_start,
+        "shoppie_aggro",
+        &StateMemory::shoppie_aggro,
+        "shoppie_aggro_next",
+        &StateMemory::shoppie_aggro_levels,
+        "merchant_aggro",
+        &StateMemory::merchant_aggro,
+        "kills_npc",
+        &StateMemory::kills_npc,
+        "level_count",
+        &StateMemory::level_count,
+        "journal_flags",
+        &StateMemory::journal_flags,
+        "time_last_level",
+        &StateMemory::time_last_level,
+        "time_level",
+        &StateMemory::time_level,
+        "level_flags",
+        &StateMemory::hud_flags,
+        "loading",
+        &StateMemory::loading,
+        "quest_flags",
+        &StateMemory::quest_flags,
+        "presence_flags",
+        &StateMemory::presence_flags,
+        "fadevalue",
+        &StateMemory::fadevalue,
+        "fadeout",
+        &StateMemory::fadeout,
+        "fadein",
+        &StateMemory::fadein,
+        "loading_black_screen_timer",
+        &StateMemory::loading_black_screen_timer,
+        "saved_dogs",
+        &StateMemory::saved_dogs,
+        "saved_cats",
+        &StateMemory::saved_cats,
+        "saved_hamsters",
+        &StateMemory::saved_hamsters,
+        "win_state",
+        &StateMemory::win_state,
+        "illumination",
+        &StateMemory::illumination,
+        "money_last_levels",
+        &StateMemory::money_last_levels,
+        "money_shop_total",
+        &StateMemory::money_shop_total,
+        "player_inputs",
+        sol::readonly(&StateMemory::player_inputs),
+        "quests",
+        &StateMemory::quests,
+        "camera",
+        &StateMemory::camera);
+    lua.new_usertype<LightParams>(
+        "LightParams",
+        "red",
+        &LightParams::red,
+        "green",
+        &LightParams::green,
+        "blue",
+        &LightParams::blue,
+        "size",
+        &LightParams::size);
+    lua.new_usertype<Illumination>(
+        "Illumination",
+        "light1",
+        &Illumination::light1,
+        "light2",
+        &Illumination::light2,
+        "light3",
+        &Illumination::light3,
+        "light4",
+        &Illumination::light4,
+        "brightness",
+        &Illumination::brightness,
+        "brightness_multiplier",
+        &Illumination::brightness_multiplier,
+        "light_pos_x",
+        &Illumination::light_pos_x,
+        "light_pos_y",
+        &Illumination::light_pos_y,
+        "offset_x",
+        &Illumination::offset_x,
+        "offset_y",
+        &Illumination::offset_y,
+        "distortion",
+        &Illumination::distortion,
+        "entity_uid",
+        &Illumination::entity_uid,
+        "flags",
+        &Illumination::flags);
+}
+}; // namespace NState
