+++ src/game_api/script/usertypes/state_lua.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <sol/forward.hpp>
+
+namespace NState
+{
+void register_usertypes(sol::state& lua);
+};
