+++ src/game_api/script/usertypes/texture_lua.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <sol/forward.hpp>
+
+class ScriptImpl;
+
+namespace NTexture
+{
+    void register_usertypes(sol::state& lua, ScriptImpl* script);
+};
